refactor: 카드 데이터 id값 추가
